requirements.txt
sqlalchemy
fastapi
uvicorn
ujson
llama_index
langchain
openai
python-dotenv